+++ Proyecto/src/main/java/com/example/Proyecto/Repository/EstadisticasNutricionalesRepository.java
@@ -13,19 +13,20 @@ public interface EstadisticasNutricionalesRepository extends JpaRepository<Estad
     // Calcular estadísticas diarias por usuario
     @Query(value = "SELECT totalCalorias, totalProteinas, totalCarbohidratos, totalGrasas, totalAzucares, totalFibra " +
             "FROM EstadisticasNutricionales WHERE idUsuario = :idUsuario AND fecha = :fecha", nativeQuery = true)
-    EstadisticasNutricionales calcularEstadisticasDiarias(@Param("idUsuario") Integer idUsuario, @Param("fecha") String fecha);
+    EstadisticasNutricionales calcularEstadisticasDiarias(@Param("idUsuario") Long idUsuario, @Param("fecha") String fecha);
 
     // Obtener progreso semanal
     @Query(value = "SELECT * FROM EstadisticasNutricionales WHERE idUsuario = :idUsuario AND fecha BETWEEN :fechaInicio AND :fechaFin ORDER BY fecha", nativeQuery = true)
-    List<EstadisticasNutricionales> obtenerProgresoSemanal(@Param("idUsuario") Integer idUsuario,
+    List<EstadisticasNutricionales> obtenerProgresoSemanal(@Param("idUsuario") Long idUsuario,
                                                            @Param("fechaInicio") String fechaInicio,
                                                            @Param("fechaFin") String fechaFin);
 
     // Calcular IMC del usuario por día
     @Query(value = "SELECT fecha, imc FROM EstadisticasNutricionales WHERE idUsuario = :idUsuario AND fecha = :fecha", nativeQuery = true)
-    Float calcularIMC(@Param("idUsuario") Integer idUsuario, @Param("fecha") String fecha);
+    Float calcularIMC(@Param("idUsuario") Long idUsuario, @Param("fecha") String fecha);
 
     // Obtener total de comidas registradas
     @Query(value = "SELECT totalComidas FROM EstadisticasNutricionales WHERE idUsuario = :idUsuario AND fecha = :fecha", nativeQuery = true)
-    Integer obtenerTotalComidasRegistradas(@Param("idUsuario") Integer idUsuario, @Param("fecha") String fecha);
+    Integer obtenerTotalComidasRegistradas(@Param("idUsuario") Long idUsuario, @Param("fecha") String fecha);
+
 }
